moved setup script to setupTests.js
