+++ src/setupTests.js
@@ -3,3 +3,35 @@
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
 import '@testing-library/jest-dom';
+import { rest } from 'msw';
+import { setupServer } from 'msw/node';
+
+const response = [
+  {
+    flickr_images: [
+      'https://imgur.com/DaCfMsj.jpg',
+      'https://imgur.com/azYafd8.jpg',
+    ],
+    name: 'Falcon 1',
+    type: 'rocket',
+    description:
+      'The Falcon 1 was an expendable launch system privately developed and manufactured by SpaceX during 2006-2009. On 28 September 2008, Falcon 1 became the first privately-developed liquid-fuel launch vehicle to go into orbit around the Earth.',
+    id: '5e9d0d95eda69955f709d1eb',
+  },
+];
+
+export const handlers = [
+  rest.get(`https://api.spacexdata.com/v3/rockets`, (req, res, ctx) => {
+    return res(ctx.json(response), ctx.delay(150));
+  }),
+];
+
+const server = setupServer(...handlers);
+// Enable API mocking before tests.
+beforeAll(() => server.listen());
+
+// Reset any runtime request handlers we may add during the tests.
+afterEach(() => server.resetHandlers());
+
+// Disable API mocking after the tests are done.
+afterAll(() => server.close());
